Update unit tests to match new conf. scores
